requirements.txt
slack_sdk
pynvml
simple-parsing
pytest-asyncio